% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confi_inter.R
\name{confi_inter}
\alias{confi_inter}
\title{Confidence Interval Function}
\usage{
confi_inter(response, covariates, alpha = 0.05)
}
\arguments{
\item{response}{A \code{vector} of response [y variable]}

\item{covariates}{A \code{matrix} of covariates [x variable]}

\item{alpha}{A \code{numeric} significance level, user input, default = 0.05}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{ci.beta}{Estimated confidence intervals}
}
}
\description{
Computes the confidence intervals of a linear regression based on user input of significance level and the approach of asymptotic.
}
\examples{
confi_inter(rep(1,10), matrix(1:20,nrow=10), alpha =  0.05)
}
\author{
Group7
}
