% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confi_inter_bo.R
\name{confi_inter_bo}
\alias{confi_inter_bo}
\title{Confidence Interval Function by Bootstrap Approach}
\usage{
confi_inter_bo(response, covariates, alpha = 0.05, m = 200)
}
\arguments{
\item{response}{A \code{vector} of response [y variable]}

\item{covariates}{A \code{matrix} of covariates[x variable]}

\item{alpha}{A \code{numeric} significance level, user input, Between 1 and 0. default = 0.05}

\item{m}{A \code{numeric} The number of bootstrap replicates, Usually this will be a single positive integer. user input, default = 200}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{ci.beta}{Estimated confidence intervals}
}
}
\description{
Computes the confidence intervals of a linear regression based on user input of significance level and the approach of bootstrap.
}
\examples{
confi_inter_bo(response, covariates, alpha= 0.05, m = 200)
}
\author{
Group7
}
