% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_test.R
\name{l_ftest}
\alias{l_ftest}
\title{F test Function}
\usage{
l_ftest(response, covariates)
}
\arguments{
\item{response}{A \code{vector} of response [y variable]}

\item{covariates}{A \code{matrix} of covariates [x variable]}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{f.value}{Calculated F value}
     \item{fp.value}{p value after comparing f calculated and f observed}
}
}
\description{
Compute F-test statistic in matrix form and the corresponding p-value
}
\examples{
l_ftest(rep(1,10), matrix(1:20,nrow=10))
}
\author{
Group7
}
