% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspe.R
\name{l_mspe}
\alias{l_mspe}
\title{Mean Square Prediction Error (MSPE) Function}
\usage{
l_mspe(response, covariates)
}
\arguments{
\item{response}{A \code{vector} dependent [y variable]}

\item{covariates}{A \code{matrix} independent [x variable]}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{mspe}{Estimated mean square prediction error}
}
}
\description{
Compute Mean square prediction error (MSPE) in matrix form where number of observations in the data is n
}
\examples{
l_mspe(rep(1,10), matrix(1:20,nrow=10))
}
\author{
Group7
}
