% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ln_plot.R
\name{ln_plot}
\alias{ln_plot}
\title{Plotting Function}
\usage{
ln_plot(response, covariates, type)
}
\arguments{
\item{response}{A \code{vector} of response [y variable]}

\item{covariates}{A \code{matrix} of covariates [x variable]}

\item{type}{A \code{list} used to control the kinds of plot. Various plots are -
RF - residuals vs fitted plot,
QQR - QQ plot of residuals,
HR - histogram of residuals,
all - all three plots}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{RF Plot}{Plot of residuals vs fitted values}
     \item{QQ Plot}{Quantile-Quantile Plot of residuals}
     \item{HR Plot}{Histogram or density of residuals plot}
}
}
\description{
Compute plots of linear regression -
- residuals vs fitted, residuals and histogram of residuals.
}
\examples{
ln_plot(rep(1,10), matrix(1:20,nrow=10),'all')
}
\author{
Group7
}
